requirements.txt
Sphinx==1.3.4
Django<2.0
dj-database-url==0.3.0
django-debug-toolbar==1.7
flake8==2.4.0
model_mommy
pytest-cov==1.8.1
pytest-django==2.8.0
pytest-sugar==0.5.1
pytest==2.7.2
psycopg2>=2.7
tox-pyenv==1.0.2
tox==2.3.1
djay==0.0.4
django-override-storage
